Log the DisplayMgr install code
The mini's NK log disappears when the DisplayMgr sets TheGDevice and
other globals. But -- the same thing happens on QEMU! So the mini is
probably not crashing in the DisplayMgr.

I used my new "patchobj" tool to make a small edit to DisplayMgr.lib.
DM_InstallDisplayManager has had its first 6 bytes replaced with a BRA.L
to an editable copy in StartInit.a